introduces raw-context-builder, on which context-builder relies. foreces resolution of library earlier.
