+++ lib/context/context-builder.js
@@ -1,42 +1,41 @@
-
-const CollectingWalker = require('../import/collecting-walker');
-const ScriptCollector = require('../import/script-collector');
-const ConfigurationCollector = require('../import/configuration-collector');
-const LibraryImporter = require('../import/lib-import/library-importer');
+const Builder = require('./raw-context-builder');
 const Providers = require('../providers/providers');
 const fs = require('fs');
 const path = require('path');
 
-/**
-  * Static utilities. Uses an internal Builder to
-  * construct the appropriate context.
-  *
-  */
-const ContextBuilder = (function(module, undefined) {
-  'use strict';
+class ContextBuilder {
 
-  let explicitRoot, explicitLib, instance;
+  constructor () {
+    this.explicitRoot = null;
+    this.explicitLib = null;
+  }
 
-  function compose(root) {
+  compose(root) {
     if (!fs.lstatSync(root).isDirectory) {
       root = path.dirname(root);
     }
-    explicitRoot = root;
-    return instance;
+    this.explicitRoot = root;
+    return this;
   }
 
-  function overrideLib(newLib) {
-    explicitLib = newLib;
-    return instance;
+  getRoot() {
+    return this.explicitRoot || getDefaultRoot();
   }
 
-  function getRoot() {
-    return explicitRoot || module.parent.filename;
+  overrideLib(newLib) {
+    this.explicitLib = resolve.call(this, newLib);
+    return this;
   }
 
-  function build(resourceName, additionalContext) {
+  getLibrary() {
+    return this.explicitLib || getDefaultLibrary.call(this, this);
+  }
+
+  build(resourceName, additionalContext) {
     const builder = new Builder()
+      .fromRoot(this.getRoot())
       .forNamedResource(resourceName)
+      .withLibrary(this.getLibrary())
       .addContext(additionalContext);
     return {
       forRest: (restConfig) => {
@@ -54,120 +53,19 @@ const ContextBuilder = (function(module, undefined) {
     };
   }
 
-  function getLibImports(scripts) {
-    new LibraryImporter(libNormalize).getImportedScripts(scripts);
-  }
+}
 
-  function normalize(name) {
-    return path.resolve(getRoot(), name);
-  }
+function getDefaultLibrary() {
+  return path.join(this.getRoot(), '..', 'lib');
+}
 
-  function libNormalize(name) {
-    if (explicitLib) {
-      let lib = path.resolve(getRoot(), explicitLib, name);
-      return lib;
-    }
-    return resolveStandardLib(name);
-  }
+function resolve(name) {
+  return path.join(this.getRoot(), name);
+}
 
-  function resolveStandardLib(name) {
-    let libRoot = path.join(getRoot(), '..', 'lib');
-    return path.resolve(libRoot, name);
-  }
+function getDefaultRoot() {
+  return module.parent.filename;
+}
 
-  function read(name) {
-    if (fs.lstatSync(name).isDirectory()) {
-      return readModule(name);
-    }
-    return readFile(name);
-  }
 
-  function readModule(moduleName) {
-    return new CollectingWalker()
-      .withDefaultOptions()
-      .withCollectors(new ScriptCollector(), new ConfigurationCollector())
-      .walk(moduleName);
-  }
-
-  function readFile(filename, errHandler) {
-    try {
-      return {
-        'scripts': [fs.readFileSync(filename, 'utf-8')]
-      };
-    } catch (err) {
-      if (errHandler) {
-        errHandler(err);
-      } else {
-        throw err;
-      }
-    }
-  }
-
-  /**
-   * Internal Builder class, allows us to defer the decision
-   * around which providers and seed to use.
-   */
-  class Builder {
-
-    constructor() {
-      this.resourceName = undefined;
-      this.additionalContext = {};
-      this.providersImpl = undefined;
-      this.seedValue = undefined;
-    }
-
-    forNamedResource(resourceName) {
-      this.resourceName = resourceName;
-      return this;
-    }
-
-    addContext(additionalContext) {
-      Object.assign(this.additionalContext, additionalContext);
-      return this;
-    }
-
-    withProvidersImpl(providersImpl) {
-      this.providersImpl = providersImpl;
-      return this;
-    }
-
-    withLocalStore(store) {
-      this.seedValue = store;
-      return this;
-    }
-
-    withRestConfig(restConfig) {
-      this.seedValue = restConfig;
-      return this;
-    }
-
-    build() {
-      const normalizedName = normalize(this.resourceName);
-      const {
-        scripts,
-        configurations
-      } = read(this.resourceName);
-      const imports = getLibImports(scripts);
-      const code = scripts
-        .concat(imports)
-        .join('\n');
-      const sandbox = Object.assign({}, {
-        Providers: this.providersImpl.seed(this.seedValue, configurations)
-      }, this.additionalContext || {});
-      return {
-        code: code,
-        sandbox: sandbox
-      };
-    }
-  }
-
-  instance = {
-    compose: compose,
-    overrideLib: overrideLib,
-    build: build
-  };
-
-  return instance;
-}());
-
-module.exports = ContextBuilder;
+module.exports = new ContextBuilder();
